+++ package/Runtime/XRCameraExtensions.cs
@@ -70,11 +70,36 @@ namespace UnityEngine.XR.ARExtensions
         /// Allows a camera provider to register support for <see cref="ICameraImageApi"/>.
         /// This is typically only used by platform-specific packages.
         /// </summary>
-        /// <param name="subsystemId"></param>
-        /// <param name="api"></param>
+        /// <param name="subsystemId">The string name associated with the camera provider to extend.</param>
+        /// <param name="api">The API to use to perform provider-specific calls.</param>
         public static void RegisterCameraImageApi(string subsystemId, ICameraImageApi api)
         {
             s_CameraImageApis[subsystemId] = api;
+        }
+
+        /// <summary>
+        /// Allows a camera provider to register support for <see cref="ICameraConfigApi"/>.
+        /// This is typically only used by platform-specific packages.
+        /// </summary>
+        /// <param name="subsystemId">The string name associated with the camera provider to extend.</param>
+        /// <param name="api">The API to use to perform provider-specific calls.</param>
+        public static void RegisterCameraConfigApi(string subsystemId, ICameraConfigApi api)
+        {
+            s_CameraConfigApis[subsystemId] = api;
+        }
+
+        /// <summary>
+        /// Allows a camera provider to register support for <see cref="TrySetFocusMode"/>.
+        /// This is typically only used by platform-specific packages.
+        /// </summary>
+        /// <param name="subsystemId">The string name associated with the camera provider to extend.</param>
+        /// <param name="handler">A method to set whether auto focus is enabled and returns <c>true</c> is successful,
+        /// or <c>false</c> otherwise.</param>
+        public static void RegisterTrySetFocusModeHandler(
+            string subsystemId,
+            Func<XRCameraSubsystem, CameraFocusMode, bool> handler)
+        {
+            s_TrySetFocusModeDelegates[subsystemId] = handler;
         }
 
         /// <summary>
@@ -126,8 +151,8 @@ namespace UnityEngine.XR.ARExtensions
         /// The <see cref="CameraImage"/> must be disposed to avoid resource leaks.
         /// </summary>
         /// <param name="cameraSubsystem">The <c>XRCameraSubsystem</c> being extended.</param>
-        /// <param name="cameraImage"></param>
-        /// <returns></returns>
+        /// <param name="cameraImage">A valid <see cref="CameraImage"/> if this method returns <c>true</c>.</param>
+        /// <returns><c>true</c> if the image was acquired; <c>false</c> otherwise.</returns>
         public static bool TryGetLatestImage(
             this XRCameraSubsystem cameraSubsystem,
             out CameraImage cameraImage)
@@ -140,9 +165,9 @@ namespace UnityEngine.XR.ARExtensions
             int planeCount;
             double timestamp;
             CameraImageFormat format;
-            if (s_AsyncCameraImageApi.TryAcquireLatestImage(out nativeHandle, out dimensions, out planeCount, out timestamp, out format))
+            if (s_CameraImageApi.TryAcquireLatestImage(out nativeHandle, out dimensions, out planeCount, out timestamp, out format))
             {
-                cameraImage = new CameraImage(s_AsyncCameraImageApi, nativeHandle, dimensions, planeCount, timestamp, format);
+                cameraImage = new CameraImage(s_CameraImageApi, nativeHandle, dimensions, planeCount, timestamp, format);
                 return true;
             }
             else
@@ -150,6 +175,91 @@ namespace UnityEngine.XR.ARExtensions
                 cameraImage = default(CameraImage);
                 return false;
             }
+        }
+
+        /// <summary>
+        /// Get the current <see cref="CameraConfiguration"/>. If camera configurations
+        /// are not supported, or if the session is not active, this method may throw
+        /// <c>InvalidOperationException</c>.
+        /// 
+        /// Camera configurations are unsupported if
+        /// <see cref="Configurations()"/> returns a collection with <see cref="CameraConfigurationCollection.count"/> of zero.
+        /// </summary>
+        /// <param name="cameraSubsystem">The <c>XRCameraSubsystem</c> being extended.</param>
+        /// <returns>The currently active <see cref="CameraConfiguration"/>.</returns>
+        public static CameraConfiguration GetCurrentConfiguration(
+            this XRCameraSubsystem cameraSubsystem)
+        {
+            if (cameraSubsystem == null)
+                throw new ArgumentNullException("cameraSubsystem");
+
+            return s_CameraConfigApi.currentConfiguration;
+        }
+
+        /// <summary>
+        /// Set a <see cref="CameraConfiguration"/> to be the active one. Throws if the
+        /// camera configuration is not supported by the device.
+        /// You can enumerate the available configurations with
+        /// <seealso cref="Configurations(XRCameraSubsystem)"/>.
+        /// 
+        /// The camera image configuration affects the resolution of the <see cref="CameraImage"/>
+        /// provided by <see cref="TryGetLatestImage(XRCameraSubsystem, out CameraImage)"/>.
+        /// It may also affect the camera framerate. See <see cref="CameraConfiguration.framerate"/>.
+        /// 
+        /// Setting this value may cause the device to relocalize, similar to stopping and restarting
+        /// the session.
+        /// 
+        /// All <see cref="CameraImage"/>s should be disposed and all asynchronous conversion operations
+        /// completed prior to setting the configuration or the this method may throw
+        /// <c>InvalidOperationException</c>, depending on the platform.
+        /// </summary>
+        /// <param name="cameraSubsystem">The <c>XRCameraSubsystem</c> being extended.</param>
+        /// <param name="configuration">A configuration to use. This is typically one of the
+        /// configurations returned by <see cref="GetConfiguration(XRCameraSubsystem, int)"/>.</param>
+        public static void SetCurrentConfiguration(
+            this XRCameraSubsystem cameraSubsystem,
+            CameraConfiguration configuration)
+        {
+            if (cameraSubsystem == null)
+                throw new ArgumentNullException("cameraSubsystem");
+
+            s_CameraConfigApi.currentConfiguration = configuration;
+        }
+
+        /// <summary>
+        /// Enumerate the <see cref="CameraConfiguration"/>s supported by the device.
+        /// </summary>
+        /// <example>
+        /// You can use <see cref="Configurations(XRCameraSubsystem)"/> in a <c>foreach</c> statement:
+        /// <code>
+        /// foreach (var config in cameraSubsystem.Configurations())
+        ///     Debug.Log(config);
+        /// </code>
+        /// </example>
+        /// <param name="cameraSubsystem">The <c>XRCameraSubsystem</c> being extended.</param>
+        /// <returns>A <see cref="CameraConfigurationCollection"/>, which allows you to
+        /// enumerate the supported configurations.</returns>
+        public static CameraConfigurationCollection Configurations(
+            this XRCameraSubsystem cameraSubsystem)
+        {
+            if (cameraSubsystem == null)
+                throw new ArgumentNullException("cameraSubsystem");
+
+            return new CameraConfigurationCollection(s_CameraConfigApi);
+        }
+
+        /// <summary>
+        /// Attempt to set the <see cref="CameraFocusMode"/>.
+        /// </summary>
+        /// <param name="cameraSubsystem">The <c>XRCameraSubsystem</c> being extended.</param>
+        /// <param name="mode">The <see cref="CameraFocusMode"/> to use.</param>
+        /// <returns><c>true</c> if the focus mode was successfully set; <c>false</c> otherwise.</returns>
+        public static bool TrySetFocusMode(this XRCameraSubsystem cameraSubsystem, CameraFocusMode mode)
+        {
+            if (cameraSubsystem == null)
+                throw new ArgumentNullException("cameraSubsystem");
+
+            return s_TrySetFocusModeDelegate(cameraSubsystem, mode);
         }
 
         /// <summary>
@@ -169,7 +279,9 @@ namespace UnityEngine.XR.ARExtensions
                 s_IsPermissionGrantedDelegate = RegistrationHelper.GetValueOrDefault(s_IsPermissionGrantedDelegates, id, DefaultIsPermissionGranted);
                 s_TryGetColorCorrectionDelegate = RegistrationHelper.GetValueOrDefault(s_TryGetColorCorrectionDelegates, id, DefaultTryGetColorCorrection);
                 s_GetNativePtrDelegate = RegistrationHelper.GetValueOrDefault(s_GetNativePtrDelegates, id, DefaultGetNativePtr);
-                s_AsyncCameraImageApi = RegistrationHelper.GetValueOrDefault(s_CameraImageApis, id, s_DefaultAsyncCameraImageApi);
+                s_CameraImageApi = RegistrationHelper.GetValueOrDefault(s_CameraImageApis, id, s_DefaultCameraImageApi);
+                s_CameraConfigApi = RegistrationHelper.GetValueOrDefault(s_CameraConfigApis, id, s_DefaultCameraConfigApi);
+                s_TrySetFocusModeDelegate = RegistrationHelper.GetValueOrDefault(s_TrySetFocusModeDelegates, id, s_DefaultTrySetFocusMode);
             }
         }
 
@@ -189,70 +301,9 @@ namespace UnityEngine.XR.ARExtensions
             return IntPtr.Zero;
         }
 
-        class DefaultCameraImageApi : ICameraImageApi
+        static bool s_DefaultTrySetFocusMode(XRCameraSubsystem cameraSubsystem, CameraFocusMode mode)
         {
-            public AsyncCameraImageConversionStatus GetAsyncRequestStatus(int requestId)
-            {
-                return AsyncCameraImageConversionStatus.Disposed;
-            }
-
-            public void DisposeImage(int nativeHandle)
-            { }
-
-            public void DisposeAsyncRequest(int requestId)
-            { }
-
-            public bool TryGetPlane(int nativeHandle, int planeIndex, out int rowStride, out int pixelStride, out IntPtr dataPtr, out int dataLength)
-            {
-                rowStride = default(int);
-                pixelStride = default(int);
-                dataPtr = default(IntPtr);
-                dataLength = default(int);
-                return false;
-            }
-
-            public bool TryAcquireLatestImage(out int nativeHandle, out Vector2Int dimensions, out int planeCount, out double timestamp, out CameraImageFormat format)
-            {
-                nativeHandle = default(int);
-                dimensions = default(Vector2Int);
-                planeCount = default(int);
-                timestamp = default(double);
-                format = default(CameraImageFormat);
-                return false;
-            }
-
-            public bool NativeHandleValid(int nativeHandle)
-            {
-                return false;
-            }
-
-            public bool TryGetConvertedDataSize(int nativeHandle, Vector2Int dimensions, TextureFormat format, out int size)
-            {
-                size = default(int);
-                return false;
-            }
-
-            public bool TryConvert(int nativeHandle, CameraImageConversionParams conversionParams, IntPtr destinationBuffer, int bufferLength)
-            {
-                return false;
-            }
-
-            public int ConvertAsync(int nativeHandle, CameraImageConversionParams conversionParams)
-            {
-                return 0;
-            }
-
-            public bool TryGetAsyncRequestData(int requestId, out IntPtr dataPtr, out int dataLength)
-            {
-                dataPtr = default(IntPtr);
-                dataLength = default(int);
-                return false;
-            }
-
-            public void ConvertAsync(int nativeHandle, CameraImageConversionParams conversionParams, OnImageRequestCompleteDelegate callback, IntPtr context)
-            {
-                callback(AsyncCameraImageConversionStatus.Disposed, conversionParams, IntPtr.Zero, 0, context);
-            }
+            return false;
         }
 
         static void SetDefaultDelegates()
@@ -260,20 +311,26 @@ namespace UnityEngine.XR.ARExtensions
             s_IsPermissionGrantedDelegate = DefaultIsPermissionGranted;
             s_TryGetColorCorrectionDelegate = DefaultTryGetColorCorrection;
             s_GetNativePtrDelegate = DefaultGetNativePtr;
-            s_AsyncCameraImageApi = s_DefaultAsyncCameraImageApi;
+            s_CameraImageApi = s_DefaultCameraImageApi;
+            s_CameraConfigApi = s_DefaultCameraConfigApi;
+            s_TrySetFocusModeDelegate = s_DefaultTrySetFocusMode;
         }
 
         static XRCameraExtensions()
         {
-            s_DefaultAsyncCameraImageApi = new DefaultCameraImageApi();
+            s_DefaultCameraImageApi = new DefaultCameraImageApi();
+            s_DefaultCameraConfigApi = new DefaultCameraConfigApi();
             SetDefaultDelegates();
         }
 
         static Func<XRCameraSubsystem, bool> s_IsPermissionGrantedDelegate;
         static TryGetColorCorrectionDelegate s_TryGetColorCorrectionDelegate;
         static Func<XRCameraSubsystem, IntPtr> s_GetNativePtrDelegate;
-        static ICameraImageApi s_AsyncCameraImageApi;
-        static DefaultCameraImageApi s_DefaultAsyncCameraImageApi;
+        static ICameraImageApi s_CameraImageApi;
+        static DefaultCameraImageApi s_DefaultCameraImageApi;
+        static ICameraConfigApi s_CameraConfigApi;
+        static DefaultCameraConfigApi s_DefaultCameraConfigApi;
+        static Func<XRCameraSubsystem, CameraFocusMode, bool> s_TrySetFocusModeDelegate;
 
         static Dictionary<string, Func<XRCameraSubsystem, bool>> s_IsPermissionGrantedDelegates =
             new Dictionary<string, Func<XRCameraSubsystem, bool>>();
@@ -286,5 +343,11 @@ namespace UnityEngine.XR.ARExtensions
 
         static Dictionary<string, ICameraImageApi> s_CameraImageApis =
             new Dictionary<string, ICameraImageApi>();
+
+        static Dictionary<string, ICameraConfigApi> s_CameraConfigApis =
+            new Dictionary<string, ICameraConfigApi>();
+
+        static Dictionary<string, Func<XRCameraSubsystem, CameraFocusMode, bool>> s_TrySetFocusModeDelegates =
+            new Dictionary<string, Func<XRCameraSubsystem, CameraFocusMode, bool>>();
     }
 }